+++ extension/js/shuffleVideo.js
@@ -16,7 +16,6 @@ async function chooseRandomVideo() {
 	const databaseSharing = configSync.databaseSharingEnabledOption;
 
 	// Get the id of the uploads playlist for this channel
-	// const uploadsPlaylistId = document.querySelector("[itemprop=channelId]").getAttribute("content").replace("UC", "UU");
 	const uploadsPlaylistId = await getPlaylistIdFromUrl(window.location.href);
 	if (!uploadsPlaylistId) {
 		throw new RandomYoutubeVideoError("Could not find channel ID. Are you on a valid page?");
@@ -40,14 +39,20 @@ async function chooseRandomVideo() {
 			playlistInfo = await getPlaylistFromApi(uploadsPlaylistId);
 
 			shouldUpdateDatabase = true;
+		} else if (databaseSharing && (playlistInfo["lastUpdatedDBAt"] ?? new Date(0).toISOString()) < addHours(new Date(), -48).toISOString()) {
+			// If the playlist exists in the database but is outdated, update it from the API.
+			console.log("Uploads playlist for this channel may be outdated in the database. Updating from the YouTube API...");
+
+			playlistInfo = await updatePlaylistFromApi(playlistInfo, uploadsPlaylistId);
+
+			shouldUpdateDatabase = true;
 		}
 
-		console.log("Uploads playlist for this channel successfully retrieved from database or API.");
+		console.log("Uploads playlist for this channel successfully retrieved.");
 
 		// The playlist exists locally, but may be outdated. Update it from the database. If needed, update the database values as well.
 	} else if ((databaseSharing && ((playlistInfo["lastFetchedFromDB"] ?? new Date(0).toISOString()) < addHours(new Date(), -48).toISOString())) ||
 		(!databaseSharing && ((playlistInfo["lastAccessedLocally"] ?? new Date(0).toISOString()) < addHours(new Date(), -48).toISOString()))) {
-
 		console.log(`Local uploads playlist for this channel may be outdated. ${databaseSharing ? "Updating from the database..." : ""}`);
 
 		playlistInfo = databaseSharing ? await tryGetPlaylistFromDB(uploadsPlaylistId) : {};
@@ -58,6 +63,7 @@ async function chooseRandomVideo() {
 			console.log(`${databaseSharing ? "Uploads playlist for this channel does not exist in the database. " : "Fetching it from the YouTube API..."}`);
 			playlistInfo = await getPlaylistFromApi(uploadsPlaylistId);
 
+			shouldUpdateDatabase = true;
 			// If the playlist exists in the database but is outdated there as well, update it from the API.
 		} else if ((playlistInfo["lastUpdatedDBAt"] ?? new Date(0).toISOString()) < addHours(new Date(), -48).toISOString()) {
 			console.log("Uploads playlist for this channel may be outdated in the database. Updating from the YouTube API...");
@@ -67,46 +73,64 @@ async function chooseRandomVideo() {
 		}
 	}
 
-	// Choose a random video from the playlist
-	let randomVideo = playlistInfo["videos"][Math.floor(Math.random() * playlistInfo["videos"].length)];
+	// Choose a random video from the videos object, where the keys are the video IDs
+	let videoIds = Object.keys(playlistInfo["videos"]) ?? [];
+	// Append the new videos to the list of videos
+	videoIds = videoIds.concat(Object.keys(playlistInfo["newVideos"] ?? {}));
+
+	let randomVideo = videoIds[Math.floor(Math.random() * videoIds.length)];
 	console.log("A random video has been chosen: " + randomVideo);
 
-	// Test if video still exists
-	let videoExists = await testVideoExistence(randomVideo);
-
+	let videoIDsToRemoveFromDB = [];
 	// If the video does not exist, remove it from the playlist and choose a new one, until we find one that exists
-	if (!videoExists) {
-		while (!videoExists) {
+	if (!await testVideoExistence(randomVideo)) {
+		do {
 			console.log("The chosen video does not exist anymore. Removing it from the database and choosing a new one...");
+			videoIDsToRemoveFromDB.push(randomVideo);
 
-			// Remove the video from the playlist
-			playlistInfo["videos"] = playlistInfo["videos"].filter(video => video !== randomVideo);
+			// Remove the video from the local playlist object
+			// It will always be in the "videos" object, as we have just fetched the "newVideos" from the YouTube API
+			delete playlistInfo["videos"][randomVideo];
 
 			// Choose a new random video
-			randomVideo = playlistInfo["videos"][Math.floor(Math.random() * playlistInfo["videos"].length)];
+			videoIds = Object.keys(playlistInfo["videos"]) ?? [];
+			// Append the new videos to the list of videos
+			videoIds = videoIds.concat(Object.keys(playlistInfo["newVideos"] ?? {}));
+
+			randomVideo = videoIds[Math.floor(Math.random() * videoIds.length)];
+
 			console.log(`A new random video has been chosen: ${randomVideo}`);
+		} while (!await testVideoExistence(randomVideo))
 
-			videoExists = await testVideoExistence(randomVideo);
-		}
-
+		// Update the database by removing the deleted videos there as well
 		shouldUpdateDatabase = true;
 	}
 
 	if (shouldUpdateDatabase && databaseSharing) {
-		console.log("Uploading the playlist to the database...");
-
 		playlistInfo["lastUpdatedDBAt"] = new Date().toISOString();
 
+		let videosToDatabase = {};
+		// If any videos need to be deleted, this should be the union of videos, newvideos, minus the videos to delete
+		if (videoIDsToRemoveFromDB.length > 0) {
+			console.log("Some videos need to be deleted from the database. All current videos will be uploaded to the database...");
+			videosToDatabase = Object.assign({}, playlistInfo["videos"], playlistInfo["newVideos"]);
+			videoIDsToRemoveFromDB.forEach(videoID => delete videosToDatabase[videoID]);
+		} else {
+			// Otherwise, we want to only upload new videos. If there are no "newVideos", we upload all videos, as this is the first time we are uploading the playlist
+			console.log("Uploading new video IDs to the database...");
+			videosToDatabase = playlistInfo["newVideos"] ?? playlistInfo["videos"] ?? {};
+		}
+
 		// Only upload the wanted keys
-		playlistInfoForDatabase = {
+		const playlistInfoForDatabase = {
 			"lastUpdatedDBAt": playlistInfo["lastUpdatedDBAt"],
 			"lastVideoPublishedAt": playlistInfo["lastVideoPublishedAt"] ?? new Date(0).toISOString(),
-			"videos": playlistInfo["videos"] ?? []
+			"videos": videosToDatabase
 		};
 
 		// Send the playlist info to the database
 		const msg = {
-			command: 'postToDB',
+			command: videoIDsToRemoveFromDB.length > 0 ? 'overwritePlaylistInfoInDB' : 'updatePlaylistInfoInDB',
 			data: {
 				key: 'uploadsPlaylists/' + uploadsPlaylistId,
 				val: playlistInfoForDatabase
@@ -119,12 +143,28 @@ async function chooseRandomVideo() {
 		playlistInfo["lastFetchedFromDB"] = playlistInfo["lastUpdatedDBAt"];
 	}
 
-	// Remember the last time the playlist was accessed locally (==now)
-	playlistInfo["lastAccessedLocally"] = new Date().toISOString();
-
 	// Update the playlist locally
 	console.log("Saving playlist to local storage...");
-	savePlaylistToLocalStorage(uploadsPlaylistId, playlistInfo);
+
+	// If applicable, add all new videos to the videos object
+	if (playlistInfo["newVideos"]) {
+		for (const videoId in playlistInfo["newVideos"]) {
+			playlistInfo["videos"][videoId] = true;
+		}
+		// remove the newVideos key
+		delete playlistInfo["newVideos"];
+	}
+
+	// Only save the wanted keys
+	playlistInfoForLocalStorage = {
+		// Remember the last time the playlist was accessed locally (==now)
+		"lastAccessedLocally": new Date().toISOString(),
+		"lastFetchedFromDB": playlistInfo["lastFetchedFromDB"] ?? new Date(0).toISOString(),
+		"lastVideoPublishedAt": playlistInfo["lastVideoPublishedAt"] ?? new Date(0).toISOString(),
+		"videos": playlistInfo["videos"] ?? {}
+	};
+
+	await savePlaylistToLocalStorage(uploadsPlaylistId, playlistInfoForLocalStorage);
 
 	// Navigate to the random video
 	window.location.href = `https://www.youtube.com/watch?v=${randomVideo}&list=${uploadsPlaylistId}`;
@@ -141,6 +181,35 @@ async function tryGetPlaylistFromDB(playlistId) {
 
 	let playlistInfo = await chrome.runtime.sendMessage(msg);
 
+	// TODO: Remove this in a future version
+	// Backward compatibility for older format where videos = [id, id] instead of videos = {id: true, id: true}
+	if (playlistInfo && playlistInfo["videos"] && Array.isArray(playlistInfo["videos"])) {
+		console.log('The information for this playlist has an old format in the database. Updating it...');
+
+		let videos = {};
+		playlistInfo["videos"].forEach(video => videos[video] = true);
+		playlistInfo["videos"] = videos;
+
+		// Update the database entry with the new format
+		// Only upload the wanted keys
+		const playlistInfoForDatabase = {
+			"lastUpdatedDBAt": playlistInfo["lastUpdatedDBAt"],
+			"lastVideoPublishedAt": playlistInfo["lastVideoPublishedAt"] ?? new Date(0).toISOString(),
+			"videos": videos
+		};
+
+		// Send the playlist info to the database
+		const msg = {
+			command: 'overwritePlaylistInfoInDB',
+			data: {
+				key: 'uploadsPlaylists/' + playlistId,
+				val: playlistInfoForDatabase
+			}
+		};
+
+		chrome.runtime.sendMessage(msg);
+	}
+
 	if (!playlistInfo) {
 		return {};
 	}
@@ -153,15 +222,13 @@ async function tryGetPlaylistFromDB(playlistId) {
 // ---------- YouTube API ----------
 
 async function getPlaylistFromApi(playlistId) {
-	let playlistInfo = {
-		"lastVideoPublishedAt": null,
-		"videos": []
-	};
+	let playlistInfo = {};
 
 	let pageToken = "";
 	let apiResponse = await getPlaylistSnippetFromAPI(playlistId, pageToken);
 
 	playlistInfo["videos"] = apiResponse["items"].map((video) => video["contentDetails"]["videoId"]);
+
 	// We also want to get the uploadTime of the most recent video
 	playlistInfo["lastVideoPublishedAt"] = apiResponse["items"][0]["contentDetails"]["videoPublishedAt"];
 	pageToken = apiResponse["nextPageToken"] ? apiResponse["nextPageToken"] : null;
@@ -173,6 +240,12 @@ async function getPlaylistFromApi(playlistId) {
 
 		pageToken = apiResponse["nextPageToken"] ? apiResponse["nextPageToken"] : null;
 	}
+
+	// Turn the videos array into an object for easier handling in the database
+	playlistInfo["videos"] = playlistInfo["videos"].reduce((obj, videoId) => {
+		obj[videoId] = true;
+		return obj;
+	}, {});
 
 	return playlistInfo;
 }
@@ -218,8 +291,11 @@ async function updatePlaylistFromApi(localPlaylist, playlistId) {
 	}
 	console.log(`Found ${newVideos.length} new video(s).`);
 
-	// Add the new videos to the localPlaylist
-	localPlaylist["videos"] = newVideos.concat(localPlaylist["videos"]);
+	// Add the new videos to a new object within the localPlaylist
+	localPlaylist["newVideos"] = newVideos.reduce((obj, videoId) => {
+		obj[videoId] = true;
+		return obj;
+	}, {});
 
 	return localPlaylist;
 }
@@ -302,6 +378,6 @@ async function tryGetPlaylistFromLocalStorage(playlistId) {
 	});
 }
 
-function savePlaylistToLocalStorage(playlistId, playlistInfo) {
-	chrome.storage.local.set({ [playlistId]: playlistInfo });
+async function savePlaylistToLocalStorage(playlistId, playlistInfo) {
+	await chrome.storage.local.set({ [playlistId]: playlistInfo });
 }